First pass at Tokenization.
Define typedef for unit tokens, which just have the type and their string
representation.

Define constructors for all token types.

Define the grand type enum for setting token and cfg terms' type field.

Add a TokenStream type and implementation, like a CharBuff but for
tokens instead.

Add a rough attempt at tokenizing a PDS file using a state machine.
The states are captured in the `tkn_sm_state` enum, and the transition
function is `tkn_sm_step`. `tokenize` threads the state machine, a file,
and a TokenStream together.