Add PDSObject & association pools.
These pools are handy for allocating a bunch of PDSObject and
Associatons structs at once and then pulling their addresses out
one-by-one if necessary. Since this program doesn't run for very long,
any sort of memory management is more trouble than it's worth, so the
only supported operations are creating a new pool and getting the
address of the next free thing in the pool. Returning objects to the
pool is not supported.

I see now why c++ templates are a thing, since the two pools are pretty
much identical except for the type of thing they contain. If the type
could be parametrized, I wouldn't need to have two basically identical
structs and sets of function implementations.