Add tokens for reserved identifiers.
Add seperate token types for the reserved identifiers `OBJECT`,
`END_OBJECT`, and `END`.

These do not add new states to the lexing state machine, it just adds
some new behavior to the identifier state. If we are about to transition
away from that state, we check the completed identifier to see if it
matches one of the reserved ones; if so, then we emit the corresponding
token rather than a generic identifier token.

Add `util.h` header file that contains the `min` utility macro.