A few small improvements to lexing.
Increase the size of the buffer used for lexing string literals to 2048
bytes.

Don't pass an argument to new_token_right_paren(), since it doesn't take
any.

Do add whitespace in a string literal when the previously inserted
character is a comma (which is also considered whitespace).

Rewind the token stream before returning it at the end of lexing, so it
can be used immediately by the caller.