Only one Assoc, but multiple Value categories.
The old PrimitiveAssoc and TupleAssoc were abstracting over tuples in
the wrong place; it's a value distinction, so the associations should
be the same. Replace these Assoc struct types with a singular
Association struct.

Value used to only encompass primitive values, but is now a union
that can be either a PrimitiveValue or a TupleValue. TupleValues contain
multiple PrimitiveValues, and PrimitiveValues are strings, dates, ints,
and floats.