Generate jar for java bindings.
The jar alone is sadly not enough, since you still need to copy the
native library to somewhere in java's `java.library.path` (apparently it
can't load native deps from within jars).

If you're in the java bindings directory, then an easy way to do this is
what's done in the Makefile's run: set the java.library.path to the
current directory.